+++ Dracula/ExternalAPIs/Kokoro/OxidationStatus.cs
@@ -1,0 +1,6 @@
+﻿namespace Shockah.Dracula;
+
+public partial interface IKokoroApi
+{
+	Status OxidationVanillaStatus { get; }
+}