+++ Dracula/docs/release-notes.md
@@ -7,6 +7,8 @@
 * Aura of Darkness B is now unplayable at 3+ Bleeding (no more infinite free Autopilot).
 * The Echolocation card now specifies the exact move it will make in its tooltip.
 * Fixed some cards using `Spr` directly, making them use completely different art in various game versions.
+* Fixed Bleeding and Blood Mirror not being affected by Timestop.
+* Fixed Blood Mirror not being marked as a "good" status.
 * Added information required to perform update checks for the mod.
 
 ## 1.1.4